+++ changelog.md
@@ -1,5 +1,11 @@
 # find-unused-exports changelog
 
+## Next
+
+### Patch
+
+- Fixed some `find-unused-exports` CLI snapshot tests.
+
 ## 1.0.0
 
 Initial release.